Removed the hidden "Related Topics" block from the sidebar.
Explicitly defined a sidebar that didn't have relations.html in it.

Changing the boolean value of show_related in html_theme_options,
too, changes the visibility of this block but it does so by only
changing its CSS 'display' property and not actually adding or
removing it from the final html.